Updated FetchNotificationInfoOperationActor
