+++ AppCore/appcore-database/src/commonMain/kotlin/me/sphere/sqldelight/operations/notifications/FetchNotificationInfoOperation.kt
@@ -1,11 +1,10 @@
 package me.sphere.sqldelight.operations.notifications
 
 import kotlinx.serialization.Serializable
-import kotlinx.serialization.builtins.serializer
 import me.sphere.sqldelight.operations.DeduplicatingInput
 import me.sphere.sqldelight.operations.OperationDefinition
 
-object FetchNotificationInfoOperation: OperationDefinition<FetchNotificationInfoOperation.Input, Unit> {
+object FetchNotificationInfoOperation: OperationDefinition<FetchNotificationInfoOperation.Input, FetchNotificationInfoOperation.Output> {
     @Serializable
     data class Input(
         val notificationId: String,
@@ -13,7 +12,12 @@ object FetchNotificationInfoOperation: OperationDefinition<FetchNotificationInfo
         override val deduplicationKey = notificationId
     }
 
+    @Serializable
+    data class Output(
+        val notificationId: String
+    )
+
     override val identifier = "FetchNotificationInfoOperation"
     override val inputSerializer = Input.serializer()
-    override val outputSerializer = Unit.serializer()
+    override val outputSerializer = Output.serializer()
 }
